requirements.txt
evdev==1.9.3
websocket-client==1.9.0
